requirements.txt
Django>=4.0.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0